[autotest] Don't pass --verify_job_repo_url on Hostless Queue Entries.
There's currently this ugly message

08/08 01:54:38.933 WARNI|subcommand:0080| parallel_simple was called
with an empty arglist, did you forget to pass in a list of machines?

in suite autoserv.DEBUG logs because we're trying to invoke the verify
job repo url control segment on hostless tasks, so we're doing
`parallel_simple(fn, [])`, which isn't accepted.

This moves it into QueueTask, which is a something that has a host.

BUG=chromium:269999
DEPLOY=scheduler
TEST=Ran a suite, and no longer see the message.  Watching scheduler
output, and suites don't get the arg, but tests do.

Change-Id: I37a56a590677893833102f99ffa242560fe6b48d
Reviewed-on: https://gerrit.chromium.org/gerrit/65183
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>