firmware_TouchMTB: Fix a bug about data_ready flag in Mtb state machine
In Mtb state machine, we collect the information about the point,
the pressure, and the syn_time. Initially, point.x and point.y are
assigned None. Their values will be updated with actual values when
the events ABS_MT_POSITION_X/Y are received. We used to have a
data_ready flag to indicate if all attributes above are non-None
values by

  data_ready = all([all(point.value()), pressure, self.syn_time])

The problem about the expression above lies in that if either X or Y
in point.value() is 0, all(point.value()) would return False while
it should be True. E.g., X may become 0 in the gesture below when the
finger draws the line to the left edge of the touchpad

  one_finger_to_edge.center_to_left

In this case, the data_ready flag should be True because 0 is the
actual value of the coordinate. In this patch, we determine if data
is ready by explicitly test that every attribute is non-None value.

BUG=chromium:235753
TEST=Run the unit test on a chromebook or in chroot to verify that
the fw strings could be extracted correctly.
(cr) $ cd .../firmware_TouchMTB
(cr) $ python tests/mtb_unittest.py
Observe that the unit test passes.

Change-Id: I12b3c659cc80a90cbc3f18583dd399c6723039c1
Reviewed-on: https://gerrit.chromium.org/gerrit/65563
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>