power_LoadTest: Make background tab last 70 minutes
We're running into errors when we try to uninstall or install the power_LoadTest
extension. It seems like we're running into a race condition between windows
closing and uninstalling/installing the extension. To work around this, make the
background window last 70 minutes instead of 60. This will ensure there is a
window to perform the uninstall of the extension on. Because of this, we don't
need to have power_LoadTest open a window at the end of the loop.

BUG=chrome-os-partner:21526
TEST=Full run of power_LoadTest

Change-Id: I51cf9d6741f9cfbc19ab6468305db2cd0fea794b
Reviewed-on: https://gerrit.chromium.org/gerrit/65433
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>