autotest: Fix site_wifitest based tests
Due a a recent change to the way SSID's are formed:

https://gerrit.chromium.org/gerrit/#/c/64955/

site_linux_router objects now raise an exception when an SSID
is requested from a router that is not currently configured.
Fix this by removing an unecessary call to disconnect the DUT
during test cleanup.  Whether the DUT is connected or not, when
profiles get popped shortly afterward, the DUT will disconnect then.

While we're here, rename _get_ssid() to _build_ssid() to distinguish it
from the more dangerous get_ssid().  The two serve very different
functions.

BUG=None
TEST=suite:wifi_matfunc passes locally.

Change-Id: Ifa9906110f59f28e191ac9b8e8ecef95fdaf6922
Reviewed-on: https://gerrit.chromium.org/gerrit/65609
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>