[autotest] add an --ssh_verbosity flag to autoserv
This CL adds an autoserv flag to tune the ssh logging verbosity. This is
accomplished by adding a global process-wide GLOBAL_SSH_COMMAND_OPTIONS
variable to site_host, to which autoserv tacks on a '-v' or '-vv' or
'-vvv' string if necessary. This is, admittedly, a little dirty. But the
make_ssh_command does not take a host object as an argument, so to add a
option host-wide to ssh_commands at the host object level would require
some major changes.

BUG=chromium:271621
TEST=Manually ran client/tests/dummy_Pass with various --ssh_verbosity
values, both valid and invalid. Saw that desired ssh verbosity was
realized.

Change-Id: If705c03f9d9d70de89642f42904e2e9fdc40bf61
Reviewed-on: https://gerrit.chromium.org/gerrit/65625
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>