Update desktopui_Respawn to handle AutoservPidAlreadyDeadError
nuke_process_by_name() will now raise AutoservPidAlreadyDeadError
if it cannot find a process by the name the caller provided.
This allows the caller to gracefully handle potential raciness when
using this method in a loop. The code here in desktopui_Respawn was
not updated to tolerate this new behavior.

BUG=chromium:273118
TEST=run the test

Change-Id: If97880b35c1e0034721cfe62221d3ec7e4a64666
Reviewed-on: https://gerrit.chromium.org/gerrit/65838
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>