Introduce a QA Validation suite.
This suite will be used by the QA team to add tests that should run
every build to give a signal around the area they are testing giving
an early indicator of potential release blocking bugs.

This is seeded with a few preliminary suites as an example. The rest
will be added over time by QA.

TEST=suite_enumerator
BUG=chromium:272999

Change-Id: I88fb6a517cfa1a908d9394b5fb3f5ca340d98b3c
Reviewed-on: https://chromium-review.googlesource.com/65839
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Alexander Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Scott Zawalski <scottz@chromium.org>