[autotest] Filter process information in drone_manager.
This change is being done as the drone appears to be filtering out
existing autoserv processes incorrectly.  It's difficult to get
debugging information out of drone_utility, so the filtering of autoserv
processes has been moved to drone_mananger.

There's also a semantic change that we now will acknowledge that there
is a process running even if it isn't an autoserv process.  This is
being done in the offchance that there's something interesting going on
with the process names at startup.

BUG=chromium:262005
DEPLOY=scheduler
TEST=unit, scheduler runs locally

Change-Id: I2413007d7c5dac79a8d26119ead639d956d6ecc6
Reviewed-on: https://gerrit.chromium.org/gerrit/65956
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>