Refactoring config settings for devices in FAFT tests.
Created a new config system for faft tests which allow for a default file of
settings and override file for each board that requires it.

The default and override files contain a class called 'Values', in which
the config attributes are stored. The class allows for inheritence of config
values from to better organize related board settings.

The new config system provides for one 'Config' class which is the main
interface to settings for all tests. It does a cascading search, for the
requested attribute, first checking a platform specific overrides, and then
through the defaults.

Modified FAFTSequence to contain one attribute 'faft_config' which replaces
self.client_attr and self.delay which it made available to tests.

BUG=None
TEST=Manual runs of various tests at desk on link device

Change-Id: I06da7be07364d5d99bdf6b204d628ad849f8d617
Reviewed-on: https://gerrit.chromium.org/gerrit/66006
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@google.com>