firmware_TouchMTB: print scores statistics with missing scores data
It is possible that some gestures have been recorded for a firmware
but are missing in the tests for another firmware. The summary module
should be able to handle this situation when printing the scores
statistics. When there are no scores data, it just leaves proper
spaces.

As an example, there are no rapid_taps_20 gesture files for fw_11.27.
Hence, it just leaves some blank spaces in the summary report.

Test Summary (by gesture) :  fw_1.0.AA           fw_11.27
                             mean    ssd count   mean    ssd count
----------------------------------------------------------------------
rapid_taps_20
  CountTrackingIDValidator:  1.00   0.00     9

BUG=chromium:235753
TEST=None. Will be tested with subsequent patches.

Change-Id: I8d3148a35008ceb6edf61a65c89f77b6c8f66fe8
Reviewed-on: https://gerrit.chromium.org/gerrit/66817
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>