firmware_TouchMTB: print metrics with missing values
It is possible that some gestures have been recorded for a firmware
but are missing in the tests for another firmware. The summary module
should be able to handle this situation when printing the metrics
statistics. When there are no metrics data, it just leaves proper
spaces.

As an example, there are no drag_edge_thumb gesture files for fw_11.27.
Hence, it just leaves some blank spaces in the summary report when
printing the metrics statistics.

Test Summary (by gesture) :  fw_1.0.AA           fw_11.27
                             mean    ssd count   mean    ssd count
-------------------------------------------------------------------------------
drag_edge_thumb
  CountTrackingIDValidator:  0.50 0.71     2

BUG=chromium:235753
TEST=None. Will be tested with subsequent patches.

Change-Id: I31311f2d00ce486082c69527b3e23ccfb3daf4f1
Reviewed-on: https://gerrit.chromium.org/gerrit/66818
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>