[autotest] Count the number of times we update an autofiled bug.
This adds a new label to autofiled bugs to indicate a count of the
number of times autofiling updates a bug.  Each time a bug is
updated, the count is incremented.  New bugs start with a count
of 1.  Pre-existing bugs without a count will be assumed to have a
count of 1 prior to increment.

BUG=chromium:265464
TEST=unit tests, plus observe bug updates to autotest-bug-filing-test from the 'dummy' suite

Change-Id: I87d9d81420f720f488e1a1bfbdefec8990b452a3
Reviewed-on: https://gerrit.chromium.org/gerrit/66430
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>