[autotest] Suite scheduling should TEST_NA unschedulable tests.
This is preperation work for dropping tryjobs.

TEST_NA'ing tests that cannot run due to DEPENDENCIES is one of the
responsibilities of the code in reimager, which we're trying to do away
with.  Now suite can handle this condition by itself, so the code in
reimager is no longer needed (and will be removed in a later CL).

BUG=chromium:250586
TEST=unit, (kicked off BVT locally with full patch stack, and create_job
failed and the tests were TEST_NA'd)

Change-Id: I92d0e00e1ee3291b1c88f353f88ee1e9b645bc4b
Reviewed-on: https://chromium-review.googlesource.com/66604
Tested-by: Alexander Miller <milleral@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Alexander Miller <milleral@chromium.org>