network_3GIdentifiers: Fix HomeProvider and ServingOperator validation logic.
The HomeProvider code is obtained from the SIM, not over-the-air, as the test
claimed. Prior to this CL, the test would fail if the network failed to provide
an operator code.
The over-the-air operator code is exposed by shill ServingOperator. This CL adds
the test to verify that modemmanager and shill consistently report this.

BUG=chromium:277823
TEST=network_3GIdentifiers passes with mu736.

Change-Id: Ia223b4253782be6488e790a656dcb23878a41d35
Reviewed-on: https://chromium-review.googlesource.com/66705
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>