FAFT: Adding 'platform' attribute to config object
Tests expect this attribute to exist. In order not to break the existing
tests, add the 'platform' attribute back. But long term its not certain
if this should be the place for this attribute, since its not config
exactly.

BUG=None
TEST=None

Change-Id: I77504f3e89830464c750aa5a15d5b17394c68deb
Reviewed-on: https://gerrit.chromium.org/gerrit/66802
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@google.com>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>