Fix Piglit.
crash_blacklist is a member of cros_ui_test and Piglit now is a simple test.
We move the blacklist to graphics_utils until we find a permanent place for
it. (But right now graphics tests are the only user of this functionality.)

BUG=chromium:271828
TEST=Ran on Alex and Link.

Change-Id: I9ae6e17d2519daa8738e0fab65941f1ac7019690
Reviewed-on: https://gerrit.chromium.org/gerrit/66862
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>