firmware_TouchMTB: save the device description file
In order to be able to replay with an existing device description
file in the test log directory, we need to save the device
description file in the test log directory in the first place.

BUG=chromium:274204
TEST=conduct the touch firmware test on a chromebook say link.
Launch the test suite. Press 'x' to exit after having recorded
a few gestures.
$ /usr/local/autotest/tests/firmware_TouchMTB
$ DISPLAY=:0 python main.py -m manual

Check the existence of "link.touchpad"
$ cat cat /var/tmp/touch_firmware_test/latest/link.touchpad

It will print something like:
N: link-touch-device
I: 0018 0000 0000 0000
P: 05 00 00 00 00 00 00 00
B: 00 0b 00 00 00 00 00 00 00
B: 01 00 00 00 00 00 00 00 00
...
A: 00 0 2040 0 0 20
A: 01 0 1360 0 0 20
A: 18 0 255 0 0 0
A: 2f 0 9 0 0 0
A: 30 0 1878 0 0 0
A: 35 0 2040 0 0 20
A: 36 0 1360 0 0 20
A: 39 0 65535 0 0 0
A: 3a 0 255 0 0 0

Also make sure that all unit tests pass.
$ python tests/run_all_unittests.py

Change-Id: I21218841250a150131e71df205c968d0e45f4c40
Reviewed-on: https://chromium-review.googlesource.com/66929
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>