[autotest] Fix suite dependencies.
Suite dependencies were not adding the dependencies to the tests that
they were scheduling, it only impacted reimaging decisions.  This adds
the labels specified in suite_dependencies to the test's DEPENDENCIES
when scheduling them.

There's also a change snuck in here to change suite_dependencies to a
list of strings, as it's incredibly frustrating trying to
programmatically manipulate a comma seperated string.

BUG=chromium:279670
TEST=unit, Ran a suite with suite dependencies.  Dependencies showed up
on scheduled test.

Change-Id: I289036499e940487592ec21e6fdbf9eb24fc4f4e
Reviewed-on: https://chromium-review.googlesource.com/167144
Tested-by: Alexander Miller <milleral@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Alexander Miller <milleral@chromium.org>