[autotest] add a nickname parameter to BgJob
This CL adds an optional nickname parameter to BgJob. If provided, the
nickname is used in any logging messages that a BgJob sends to the
logging manager. The parameter is only specified by the
master-ssh-connection call to BgJob at the moment, but in a future CL
that will be extended to more BgJobs.

BUG=None
TEST=None

Change-Id: Iee7722b55cc5e7497888ee86154c3ddb3b7bdf6e
Reviewed-on: https://chromium-review.googlesource.com/167172
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>