[autotest] All suites now specify version_prefix.
The Reimager is how we know what version prefix to use.  We now have the
suite specify the version prefix to use, so that we don't rely on the
reimager.  This then gets threaded into the suite_dependencies so that
it gets attached onto all tests before we go to schedule them.

devserver_url was also threaded through in a couple suite control files
that were missing it.

BUG=chromium:250879, chromium:250586
TEST=Made trybot build trybot-lumpy-paladin/R31-4602.0.0-b1134 with
control file changes.  Ran a suite locally and verified suites are still
scheduled as expected.  (And double DEPENDENCIES are de-deuplicated.)

Change-Id: I00e06108dd5621086802bf7a7d447327d1f2ca72
Reviewed-on: https://chromium-review.googlesource.com/66605
Tested-by: Alexander Miller <milleral@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Alexander Miller <milleral@chromium.org>