[autotest] Throttle HostlessQueueEntries.
This is a quick hack to throttling the number of suites we have active.
I'd like if there was a nice easy way to single out what a suite is, or
have a job elect to be throttled, but in terms of making sure the system
doesn't go down again soon for the same reason, this will be good enough
for now.

Note that, while throttled, the suite jobs show as running, but the
autoserv process isn't kicked off.

BUG=chromium:279627
DEPLOY=scheduler
TEST=Kicked off multiple suites with max_hostless_processes set to 500.
All ran.  Set max_hostless_processes to 1, and reloaded via :13467.
Scheduled another suite, which didn't start running until all other
suites finished.

Change-Id: I141a9bc4bd64d9a40c1ec4cd05a624739bb85076
Reviewed-on: https://chromium-review.googlesource.com/167175
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>