firmware_TouchMTB: flush logs without mtplot images
Sometimes, it is handy to be able to replay the logs without mtplot
images. This may be required when the mtplot images were not captured
correctly due to some other problems.

When flushing the test results to the html result file, we only
check the existence of the gesture logs and not the image files.

BUG=chromium:274204
TEST=Replay the existing logs on a chromebook machine after removing png files.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ rm tests/logs/link/20130806_221321-fw_1.0.AA-robot/*.png
$ tools/machine_replay.sh -b tests/logs/link -s
Observe that the replay could finish correctly.

Change-Id: I708f492c4f90a115fec61a15c5f3a628d9aea5e8
Reviewed-on: https://chromium-review.googlesource.com/167452
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>