firmware_TouchMTB: chop off pauses at both ends in tracking gestures
When the touch robot or a tester performs tracking gestures,
occasionally, there are pauses in the beginning and in the end.
While the kernel only reports changes of MTB attributes such as
x, y, and pressure values, a packet with the same attributes as
its previous one will be simply discarded in the kernel which
leads to larger packet report intervals.

In this patch, only the middle segment in which points are away
from both ends by more than a specified distance are used to
calculate the report rate metrics. In other words, the points
at both ends are chopped off in order not to generate the false
alarms of large packet report intervals.

BUG=chromium:252152
TEST=Conduct unit tests on a chromebook or in chroot.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ python tests/run_all_unittests.py
The unit tests must all pass.

Change-Id: Ia2eb64a5b86d81a5125948d3f87b0705836c6f1f
Reviewed-on: https://chromium-review.googlesource.com/168908
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>