Add a test for OBSS background scan back-off
This test is similar to the current BgscanBackoff test, except that
scans are triggered by the 802.11n Overlapping BSS detection scans
as opposed to wpa_supplicant's background scan mechanism.  To do so,
we add configuration to wpa_supplicant to request that clients perform
these scans, then test both that the client performs the scan (by
monitoring the hostapd log) and that these scans do not disrupt the
latency.

CQ-DEPEND=CL:167975
BUG=chrome-os-partner:22474
TEST=This is a test

Change-Id: I0b4d2393f35637cf37670f79afc8e40a82f6c256
Reviewed-on: https://chromium-review.googlesource.com/168122
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>