[autotest] improve abstract_ssh's detection of a down master-ssh
If the master-ssh connection closes itself, it appears that it will
remove its ControlPath socket file. Based on the logs linked in the bug,
this can happen without abstract_ssh noticing.

This CL fixes that by checking for the existence of the ControlPath
socket file, in addition to the existing master_ssh_job.sp.poll() test.

BUG=chromium:286014
TEST=Tryjobs:
http://chromegw/i/chromiumos.tryserver/builders/lumpy-paladin/builds/1145
http://chromegw/i/chromiumos.tryserver/builders/mario-paladin/builds/2106

Change-Id: Ic6e6b83378b7f5aa3280bbae4c5946584cd69df6
Reviewed-on: https://chromium-review.googlesource.com/168302
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>