[autotest] Set timeout and priority based on frequency of scheduling.
Out-of-band suites now have variable priority and duration based on the
event that kicks off the task.

Event     | Priority  | Duration
new_build | PostBuild | 12 hours
nightly   | Daily     | One day
weekly    | Weekly    | One week

This will mean that out-of-band suites will have lower priority than
user-created jobs in the lab (which are made with the Default priority)
and much lower than the inline build processes.  The ability for a
Weekly suite to live for a whole week is to compensate for this drop in
priority.

I expect that this will fix a lot of the problems we have with
overlapping pools.  Now if we overlapped pool:bvt with pool:suites, we
wouldn't see out-of-band suites starving the BVT out of machines, as BVT
would have a higher priority and thus preempt the out-of-band suites.

BUG=chromium:221264
DEPLOY=suite_scheduler
TEST=unit, force build and saw timeouts and priorities appear correctly

Change-Id: I604b7869d7fe16859f09a9effb11cbcb33980e99
Reviewed-on: https://chromium-review.googlesource.com/168569
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>