[autotest] `Clone job` is currently broken on the AFE. This unbreaks.
I missed a case where priority was still expected to be a string. :(

There's also a fix in here that if you clone a job with a higher
priority than what you're allowed to schedule, it's automatically bumped
down to the maximum of what you can schedule.

BUG=None
DEPLOY=afe
TEST=cloned job

Change-Id: Ie00531ea478379b753c6a3b3bbbb1cac82da040e
Reviewed-on: https://chromium-review.googlesource.com/168560
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>