[AU] Always update the devserver code when importing it.
If there are dependencies where ToT autotest needs something from ToT
devserver, then one needs to make sure that the devserver code is
updated before importing it.

This is doubly important in the context of the builders, as they seem to
clone the devserver code once, and then never update it.

BUG=None
TEST=Ran a full_release_test.py that I copied from AUTest step.
Devserver code was updated.

Change-Id: Iffd1a2dd3561cd75d3d495460619515e011d3547
Reviewed-on: https://chromium-review.googlesource.com/168551
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>