FAFT: Set the shutdown confirmation period to a small value larger than 0
We original set the shutdown confirmation period to 0 to speed-to the check,
meaning that it don't wait 60s (the default value) to confirm DUT shutdown.

After we refactored to use the AbstractSshHost class to handle the connection,
some method treats the value 0 as an error and it also doesn't make sense.
So this change is to make it a small value larger than 0.

BUG=chrome-os-partner:21118;chromium-os:215491
TEST=Manual
Ran the firmware_FwScreenCloseLid test passed on Spring.

Change-Id: I91bfef922b9daa035168c4ea8835fa0838985734
Reviewed-on: https://chromium-review.googlesource.com/168543
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>