firmware_TouchMTB: add scripts to print SYN_REPORT intervals
It is handy to have scripts to print the time intervals between
SYN_REPORT packets in real time for a touchpad or touchscreen
device. In addition, the number of packets with large intervals
greater than 1/60 second is also displayed. A copy of the raw data
is saved with the filename specified by a command line option.

BUG=chromium:252152
TEST=Execute the script at any directory on a chromebook.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
The default device node is touchpad.
$ tools/print_syn.sh

Or you can specify 'touchscreen' on a Pixel.
$ tools/print_syn.sh -d touchscreen

It would print something like:

        ...
        interval: 0.008986   (Event: time 273574.295486, --- SYN_REPORT ---)
        interval: 0.009091   (Event: time 273574.304577, --- SYN_REPORT ---)
        interval: 0.008975   (Event: time 273574.313552, --- SYN_REPORT ---)
  ***** interval: 0.034414   (Event: time 273574.347966, --- SYN_REPORT ---)
        ...

the count of large intervals: 1 out of 23 packets

You could also specify the raw data file name with the "-f" option. E.g.,
$ tools/print_syn.sh -f /tmp/circle_interval.dat
And check the existence of the raw data file.

Change-Id: Ib1c62ae41b907e8ff7bf0d7a5bb43c895c9de443
Reviewed-on: https://chromium-review.googlesource.com/168425
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>